+++ lib/index.ts
@@ -1,11 +1,13 @@
 'use strict';
-import type { ConnectOptions, CreateLocalTrackOptions } from '../tsdef/types';
+
+import type { ConnectOptions, CreateLocalTrackOptions, CreateLocalAudioTrackOptions } from '../tsdef/types';
 import type { LocalAudioTrack as LocalAudioTrackType } from '../tsdef/LocalAudioTrack';
 import type { LocalVideoTrack as LocalVideoTrackType } from '../tsdef/LocalVideoTrack';
 import type { Log } from '../tsdef/loglevel';
 import type { Room } from '../tsdef/Room';
 import { createLocalTracks } from './createlocaltracks';
 import { runPreflight } from './preflight/preflighttest';
+
 
 const internals = {
   connect: require('./connect'),
@@ -27,7 +29,7 @@ function connect(token: string, options?: ConnectOptions): Promise<Room> {
   return internals.connect(token, internalOptions);
 }
 
-function createLocalAudioTrack(options?: CreateLocalTrackOptions): Promise<LocalAudioTrackType> {
+function createLocalAudioTrack(options?: CreateLocalTrackOptions|CreateLocalAudioTrackOptions): Promise<LocalAudioTrackType> {
   const internalOptions = {
     createLocalTracks,
     ...options
@@ -77,5 +79,5 @@ module.exports = {
   Logger,
   LocalAudioTrack,
   LocalVideoTrack,
-  LocalDataTrack
+  LocalDataTrack,
 };