+++ tsdef/LocalAudioTrack.d.ts
@@ -1,11 +1,13 @@
 import { AudioTrack } from './AudioTrack';
 import { LocalTrackOptions } from './LocalTrackOptions';
+import { NoiseCancellation } from './types';
 
 export class LocalAudioTrack extends AudioTrack {
   constructor(mediaStreamTrack: MediaStreamTrack, options?: LocalTrackOptions);
 
   isStopped: boolean;
   mediaStreamTrack: MediaStreamTrack;
+  noiseCancellation?: NoiseCancellation;
 
   disable(): this;
   enable(enabled?: boolean): this;