+++ tsdef/LocalTrackOptions.d.ts
@@ -1,4 +1,4 @@
 export interface LocalTrackOptions {
-    name?: string;
-    workaroundWebKitBug1208516?: boolean
-  }
+  name?: string;
+  workaroundWebKitBug1208516?: boolean
+}